# storage/swagger/http.json
{"openapi":"3.0.0","info":{"title":"Saque PIX API","description":"API para saques via PIX com notificações por email","version":"1.0.0"},"servers":[{"url":"http:\/\/127.0.0.1:9502","description":"Saque PIX API Server"}],"paths":{"\/admin\/process-scheduled-withdraws":{"post":{"tags":["Admin"],"summary":"Processar saques agendados manualmente","description":"Dispara manualmente o processamento de saques agendados (simula a cron).","operationId":"processScheduledWithdraws","responses":{"200":{"description":"Processamento concluído","content":{"application\/json":{"schema":{"properties":{"message":{"type":"string","example":"Processamento de saques agendados concluído","rules":null,"attribute":null}},"type":"object"}}}}}}},"\/health":{"get":{"tags":["Health"],"summary":"Health check","description":"Verifica status da API e conexão com banco de dados","operationId":"d31d64e2487861d1663db2aa02aa30b5","responses":{"200":{"description":"API funcionando normalmente","content":{"application\/json":{"schema":{"properties":{"status":{"type":"string","example":"ok"},"timestamp":{"type":"string","example":"2026-05-13 15:30:00"},"database":{"properties":{"status":{"type":"string","example":"ok"},"message":{"type":"string","example":"Connected"}},"type":"object"}},"type":"object"}}}}}}},"\/health\/error":{"get":{"tags":["Health"],"summary":"Testar erro 500","description":"Força um erro de servidor para testar o handler de exceções","operationId":"cf0c5248dd0bab375fd6fcda606f029b","responses":{"500":{"description":"Erro de servidor","content":{"application\/json":{"schema":{"properties":{"code":{"type":"integer","example":500},"message":{"type":"string","example":"Server Error"},"errors":{"type":"object"}},"type":"object"}}}}}}},"\/health\/validation-error":{"get":{"tags":["Health"],"summary":"Testar erro 422","description":"Força um erro de validação para testar o handler de validação","operationId":"df478e84b0426eace0c2706b11e57f8e","responses":{"422":{"description":"Erro de validação","content":{"application\/json":{"schema":{"properties":{"code":{"type":"integer","example":422},"message":{"type":"string","example":"Validation Error"},"errors":{"type":"object","example":{"amount":["Amount must be greater than 0"]}}},"type":"object"}}}}}}},"\/health\/test-email":{"post":{"tags":["Health"],"summary":"Testar envio de email","description":"Envia um email de teste usando o template de saque PIX","operationId":"35a9f5bd82cd975f428671b7897249e5","requestBody":{"required":true,"content":{"application\/json":{"schema":{"properties":{"email":{"description":"Endereço de email do destinatário","type":"string","format":"email","example":"teste@exemplo.com"}},"type":"object"}}}},"responses":{"200":{"description":"Email enviado com sucesso","content":{"application\/json":{"schema":{"properties":{"code":{"type":"integer","example":200},"message":{"type":"string","example":"Test email sent successfully"},"data":{"properties":{"to":{"type":"string","example":"teste@exemplo.com"},"check_mailpit":{"type":"string","example":"http:\/\/localhost:8025"}},"type":"object"}},"type":"object"}}}},"400":{"description":"Email não fornecido","content":{"application\/json":{"schema":{"properties":{"code":{"type":"integer","example":400},"message":{"type":"string","example":"Email is required"}},"type":"object"}}}},"500":{"description":"Falha ao enviar email","content":{"application\/json":{"schema":{"properties":{"code":{"type":"integer","example":500},"message":{"type":"string","example":"Failed to send test email"}},"type":"object"}}}}}}},"\/account\/{accountId}\/balance\/withdraw":{"post":{"tags":["Saque"],"summary":"Realizar saque PIX","description":"Saque imediato ou agendado via PIX. Para saque imediato passe schedule como null.","operationId":"withdraw","parameters":[{"name":"accountId","in":"path","description":"UUID da conta","required":true,"schema":{"type":"string","format":"uuid","example":"69770a98-38d4-4956-b996-f40657116cdc"}}],"requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["method","pix","amount"],"properties":{"method":{"type":"string","enum":["PIX"],"example":"PIX","rules":null,"attribute":null},"pix":{"required":["type","key"],"properties":{"type":{"type":"string","enum":["email"],"example":"email","rules":null,"attribute":null},"key":{"type":"string","example":"usuario@email.com","rules":null,"attribute":null}},"type":"object","rules":null,"attribute":null},"amount":{"type":"number","format":"double","example":150.75,"rules":null,"attribute":null},"schedule":{"type":"string","format":"date-time","example":null,"nullable":true,"rules":null,"attribute":null}},"type":"object"}}},"_content":{"required":["method","pix","amount"],"properties":{"method":{"type":"string","enum":["PIX"],"example":"PIX","rules":null,"attribute":null},"pix":{"required":["type","key"],"properties":{"type":{"type":"string","enum":["email"],"example":"email","rules":null,"attribute":null},"key":{"type":"string","example":"usuario@email.com","rules":null,"attribute":null}},"type":"object","rules":null,"attribute":null},"amount":{"type":"number","format":"double","example":150.75,"rules":null,"attribute":null},"schedule":{"type":"string","format":"date-time","example":null,"nullable":true,"rules":null,"attribute":null}},"type":"object"}},"responses":{"200":{"description":"Saque realizado ou agendado com sucesso","content":{"application\/json":{"schema":{"properties":{"message":{"type":"string","example":"Saque realizado com sucesso","rules":null,"attribute":null},"withdraw_id":{"type":"string","format":"uuid","rules":null,"attribute":null},"status":{"type":"string","example":"completed","rules":null,"attribute":null},"amount":{"type":"number","format":"double","example":150.75,"rules":null,"attribute":null}},"type":"object"}}}},"400":{"description":"Erro de validação ou saldo insuficiente","content":{"application\/json":{"schema":{"properties":{"message":{"type":"string","example":"Saldo insuficiente","rules":null,"attribute":null},"code":{"type":"integer","example":400,"rules":null,"attribute":null}},"type":"object"}}}},"404":{"description":"Conta não encontrada","content":{"application\/json":{"schema":{"properties":{"message":{"type":"string","example":"Conta não encontrada","rules":null,"attribute":null},"code":{"type":"integer","example":404,"rules":null,"attribute":null}},"type":"object"}}}}}}}},"tags":[{"name":"Admin","description":"Admin"},{"name":"Health","description":"Health"},{"name":"Saque","description":"Saque"}]}